made some refactoring and cleaning up
